Fixed the form field

+++ workshift/forms.py
@@ -390,7 +390,8 @@ class SignOutForm(InteractShiftForm):
 
 class AddWorkshifterForm(forms.Form):
 	add_profile = forms.BooleanField(initial=True)
-	hours = forms.IntegerField(min_value=0, initial=settings.DEFAULT_SEMESTER_HOURS)
+	hours = forms.DecimalField(min_value=0, max_digits=7, decimal_places=2,
+							   initial=settings.DEFAULT_SEMESTER_HOURS)
 
 	def __init__(self, *args, **kwargs):
 		self.semester = kwargs.pop("semester")
